Simplify the Tennis kata to just one version, java
