--- 04-TennisScoring/java/src/main/java/ITennisGame.java
@@ -1,8 +1,0 @@
-
-public interface ITennisGame {
-
-  String getScore();
-
-  void wonPoint(String player);
-
-}
