+++ 04-TennisScoring/java/src/main/java/TennisGame.java
@@ -1,12 +1,12 @@
 
-public class TennisGameMedium implements ITennisGame {
-    
+public class TennisGame {
+
     private int m_score1 = 0;
     private int m_score2 = 0;
     private String player1Name;
     private String player2Name;
 
-    public TennisGameMedium(String player1Name, String player2Name) {
+    public TennisGame(String player1Name, String player2Name) {
         this.player1Name = player1Name;
         this.player2Name = player2Name;
     }
@@ -37,7 +37,7 @@ public class TennisGameMedium implements ITennisGame {
                 default:
                         score = "Deuce";
                     break;
-                
+
             }
         }
         else if (m_score1>=4 || m_score2>=4)
